Showing the upload percentage for assets on table
