+++ packages/www/components/Dashboard/AssetsTable/helpers.tsx
@@ -1,6 +1,5 @@
 import { Asset, Task } from "@livepeer.studio/api";
 import { Box } from "@livepeer/design-system";
-import { FileUpload, FileUploadsDictionary } from "hooks/use-api";
 import { get } from "lodash";
 import ActionCell, { ActionCellProps } from "../Table/cells/action";
 import CreatedAtCell, { CreatedAtCellProps } from "../Table/cells/createdAt";
@@ -63,23 +62,11 @@ export type RowsPageFromStateResult = {
   count: any;
 };
 
-const fileUploadProgressForAsset = (
-  asset: Asset,
-  fileUploads: FileUpload[]
-): number | undefined => {
-  const fileUpload = fileUploads.find(
-    (upload) =>
-      asset.name === upload.file.name && asset.status?.phase === "waiting"
-  );
-  return fileUpload ? fileUpload.progress : undefined;
-};
-
 export const rowsPageFromState = async (
   state,
   userId: string,
   getAssets: Function,
   getTasks: Function,
-  currentFileUploads: FileUploadsDictionary,
   onDeleteAsset: Function
 ): Promise<RowsPageFromStateResult> => {
   const [assets, nextCursor, count] = await getAssets(userId, {
@@ -92,10 +79,6 @@ export const rowsPageFromState = async (
 
   const [tasks] = await getTasks(userId);
 
-  const fileUploadsFiltered = Object.keys(currentFileUploads ?? {})
-    .map((key) => currentFileUploads?.[key])
-    .filter((file) => file && !file.error && file.file.name && !file.completed);
-
   const rows: AssetsTableData[] = assets.map(
     (asset: Asset): AssetsTableData => {
       const source: Task =
@@ -104,12 +87,6 @@ export const rowsPageFromState = async (
 
       const isStatusFailed = asset.status.phase === "failed";
       const { errorMessage } = asset.status;
-
-      const fileUploadProgress = fileUploadProgressForAsset(
-        asset,
-        fileUploadsFiltered
-      );
-      const isFileUploading = fileUploadProgress !== undefined;
 
       return {
         id: asset.id,
@@ -138,10 +115,9 @@ export const rowsPageFromState = async (
           date: new Date(asset.createdAt),
           fallback: <Box css={{ color: "$primary8" }}>—</Box>,
           href: `/dashboard/assets/${asset.id}`,
+          asset,
           isStatusFailed,
           errorMessage,
-          isFileUploading,
-          fileUploadProgress,
         },
         updatedAt: {
           date:
