+++ packages/www/components/Dashboard/Table/cells/createdAt.tsx
@@ -1,6 +1,10 @@
+import { Asset } from "@livepeer.studio/api";
 import { Badge, Flex } from "@livepeer/design-system";
 import { UploadIcon } from "@radix-ui/react-icons";
 import { format } from "date-fns";
+import { useApi } from "hooks";
+import { FileUpload, FileUploadsDictionary } from "hooks/use-api";
+import { useEffect, useMemo, useState } from "react";
 import ReactTooltip from "react-tooltip";
 import { CellComponentProps, TableData } from "../types";
 
@@ -45,6 +49,17 @@ const FileUploading = ({ progress }) => (
   </Flex>
 );
 
+const fileUploadProgressForAsset = (
+  asset: Asset,
+  fileUploads: FileUpload[]
+): number | undefined => {
+  const fileUpload = fileUploads.find(
+    (upload) =>
+      asset.name === upload.file.name && asset.status?.phase === "waiting"
+  );
+  return fileUpload ? fileUpload.progress : undefined;
+};
+
 export type CreatedAtCellProps = {
   id: string;
   date: Date;
@@ -52,22 +67,34 @@ export type CreatedAtCellProps = {
   href?: string;
   isStatusFailed: boolean;
   errorMessage?: string;
-  isFileUploading: boolean;
-  fileUploadProgress?: number;
+  asset: Asset;
 };
 
 const CreatedAtCell = <D extends TableData>({
   cell,
 }: CellComponentProps<D, CreatedAtCellProps>) => {
-  const {
-    id,
-    date,
-    fallback,
-    isStatusFailed,
-    errorMessage,
-    isFileUploading,
-    fileUploadProgress,
-  } = cell.value;
+  const { id, date, fallback, asset, isStatusFailed, errorMessage } =
+    cell.value;
+
+  const { getCurrentFileUploads } = useApi();
+  const [currentFileUploads, setCurrentFileUploads] = useState<
+    FileUploadsDictionary | undefined
+  >();
+  const fileUploadProgress = useMemo(() => {
+    const fileUploadsFiltered = Object.keys(currentFileUploads ?? {})
+      .map((key) => currentFileUploads?.[key])
+      .filter(
+        (file) => file && !file.error && file.file.name && !file.completed
+      );
+    return fileUploadProgressForAsset(asset, fileUploadsFiltered);
+  }, [currentFileUploads]);
+  const isFileUploading = fileUploadProgress !== undefined;
+
+  useEffect(() => {
+    // Has to fetch the file uploads from this cell, instead of from the table fetcher, to avoid fetching again assets too
+    const fileUploads = getCurrentFileUploads();
+    setCurrentFileUploads(fileUploads);
+  });
 
   if (isStatusFailed) {
     return <FailedProcessing id={id} errorMessage={errorMessage} />;
